requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
requests
beautifulsoup4
python-dotenv